test init migration to db

+++ TSM.Task.Domain/Entities/Task.cs
@@ -9,9 +9,9 @@ public class Task
 	public DateTime? Dedline { get; set; }
 	public string? Comment { get; set; }
 
-	public int CategoryId { get; set; }
+	public byte CategoryId { get; set; }
 	public Category? Category { get; set; }
 
-	public int PriorityId { get; set; }
+	public byte PriorityId { get; set; }
 	public Priority? Priority { get; set; }
 }
